IMPALA-13256: Support more than 2G rows for COUNT(*) on jdbc table
Function DatabaseAccessor.getTotalNumberOfRecords() is defined with
return data type as int. This caused Impala to throw an exception when
executing COUNT(*) query for jdbc tables with more than 2G rows.
This patch fixed the issue by changing the function return type as
long. It also made number of rows in each TRowBatch fetched from jdbc
data source not to exceed 2G.

Testing:
 - Passed core-test and tpcds test for jdbc tables.
 - Manually created a jdbc table jdbc_table with more than 2G rows,
   verified that query 'select count(*) from jdbc_table' returned
   correct number of rows. Detailed steps were added in the comments
   of IMPALA-13256.

Reviewed-on: http://gerrit.cloudera.org:8080/21617
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>

+++ fe/src/main/java/org/apache/impala/extdatasource/jdbc/dao/DatabaseAccessor.java
@@ -24,7 +24,7 @@ import org.apache.impala.extdatasource.jdbc.exception.JdbcDatabaseAccessExceptio
 
 public interface DatabaseAccessor {
 
-  int getTotalNumberOfRecords(Configuration conf)
+  long getTotalNumberOfRecords(Configuration conf)
       throws JdbcDatabaseAccessException;
 
   JdbcRecordIterator getRecordIterator(Configuration conf, int limit, int offset)
